+++ app/upgrades/v20/upgrades.go
@@ -19,7 +19,6 @@ func CreateUpgradeHandler(
 	return func(c context.Context, _ upgradetypes.Plan, vm module.VersionMap) (module.VersionMap, error) {
 		ctx := sdk.UnwrapSDKContext(c)
 		logger := ctx.Logger().With("upgrade", UpgradeName)
-
 		logger.Debug("Running module migrations...")
 		return mm.RunMigrations(ctx, configurator, vm)
 	}