+++ x/evm/keeper/keeper_test.go
@@ -1,13 +1,11 @@
 package keeper_test
 
 import (
-	"fmt"
 	"math/big"
 
 	sdk "github.com/cosmos/cosmos-sdk/types"
-	authtypes "github.com/cosmos/cosmos-sdk/x/auth/types"
 
-	"github.com/evmos/evmos/v19/utils"
+	evmostypes "github.com/evmos/evmos/v19/types"
 	"github.com/evmos/evmos/v19/x/evm/keeper"
 	"github.com/evmos/evmos/v19/x/evm/statedb"
 	evmtypes "github.com/evmos/evmos/v19/x/evm/types"
@@ -97,15 +95,14 @@ func (suite *KeeperTestSuite) TestGetAccountStorage() {
 		malleate func() common.Address
 	}{
 		{
-			name:     "Only accounts that are not a contract (no storage)",
+			name:     "Only one account that's not a contract (no storage)",
 			malleate: nil,
 		},
 		{
-			name: "One contract (with storage) and other EOAs",
+			name: "Two accounts - one contract (with storage), one wallet",
 			malleate: func() common.Address {
 				supply := big.NewInt(100)
-				contractAddr := suite.DeployTestContract(suite.T(), ctx, suite.keyring.GetAddr(0), supply)
-				return contractAddr
+				return suite.DeployTestContract(suite.T(), ctx, suite.keyring.GetAddr(0), supply)
 			},
 		},
 	}
@@ -121,23 +118,17 @@ func (suite *KeeperTestSuite) TestGetAccountStorage() {
 			}
 
 			i := 0
-			suite.network.App.AccountKeeper.IterateAccounts(ctx, func(account sdk.AccountI) bool {
-				acc, ok := account.(*authtypes.BaseAccount)
+			suite.network.App.AccountKeeper.IterateAccounts(suite.network.GetContext(), func(account sdk.AccountI) bool {
+				ethAccount, ok := account.(evmostypes.EthAccountI)
 				if !ok {
-					// Ignore e.g. module accounts
+					// ignore non EthAccounts
 					return false
 				}
 
-				address, err := utils.Bech32ToHexAddr(acc.Address)
-				if err != nil {
-					// NOTE: we panic in the test to see any potential problems
-					// instead of skipping to the next account
-					panic(fmt.Sprintf("failed to convert %s to hex address", err))
-				}
+				addr := ethAccount.EthAddress()
+				storage := suite.network.App.EvmKeeper.GetAccountStorage(suite.network.GetContext(), addr)
 
-				storage := suite.network.App.EvmKeeper.GetAccountStorage(ctx, address)
-
-				if address == contractAddr {
+				if addr.Hex() == contractAddr.Hex() {
 					suite.Require().NotEqual(0, len(storage),
 						"expected account %d to have non-zero amount of storage slots, got %d",
 						i, len(storage),
@@ -148,7 +139,6 @@ func (suite *KeeperTestSuite) TestGetAccountStorage() {
 						i, 0, len(storage),
 					)
 				}
-
 				i++
 				return false
 			})