+++ x/evm/keeper/statedb.go
@@ -1,18 +1,19 @@
 // Copyright Tharsis Labs Ltd.(Evmos)
 // SPDX-License-Identifier:ENCL-1.0(https://github.com/evmos/evmos/blob/main/LICENSE)
-
 package keeper
 
 import (
-	"errors"
+	"fmt"
 	"math/big"
 
+	errorsmod "cosmossdk.io/errors"
 	sdkmath "cosmossdk.io/math"
 
 	"cosmossdk.io/store/prefix"
 	storetypes "cosmossdk.io/store/types"
 	sdk "github.com/cosmos/cosmos-sdk/types"
 	"github.com/ethereum/go-ethereum/common"
+	evmostypes "github.com/evmos/evmos/v19/types"
 	"github.com/evmos/evmos/v19/x/evm/statedb"
 	"github.com/evmos/evmos/v19/x/evm/types"
 )
@@ -23,7 +24,7 @@ var _ statedb.Keeper = &Keeper{}
 // StateDB Keeper implementation
 // ----------------------------------------------------------------------------
 
-// GetAccount returns nil if account is not exist
+// GetAccount returns nil if account is not exist, returns error if it's not `EthAccountI`
 func (k *Keeper) GetAccount(ctx sdk.Context, addr common.Address) *statedb.Account {
 	acct := k.GetAccountWithoutBalance(ctx, addr)
 	if acct == nil {
@@ -44,43 +45,6 @@ func (k *Keeper) GetState(ctx sdk.Context, addr common.Address, key common.Hash)
 	}
 
 	return common.BytesToHash(value)
-}
-
-// GetFastState loads contract state from database.
-func (k *Keeper) GetFastState(ctx sdk.Context, addr common.Address, key common.Hash) []byte {
-	store := prefix.NewStore(ctx.KVStore(k.storeKey), types.AddressStoragePrefix(addr))
-
-	return store.Get(key.Bytes())
-}
-
-// GetCodeHash loads the code hash from the database for the given contract address.
-func (k *Keeper) GetCodeHash(ctx sdk.Context, addr common.Address) common.Hash {
-	store := prefix.NewStore(ctx.KVStore(k.storeKey), types.KeyPrefixCodeHash)
-	bz := store.Get(addr.Bytes())
-	if len(bz) == 0 {
-		return common.BytesToHash(types.EmptyCodeHash)
-	}
-
-	return common.BytesToHash(bz)
-}
-
-// IterateContracts iterates over all smart contract addresses in the EVM keeper and
-// performs a callback function.
-//
-// The iteration is stopped when the callback function returns true.
-func (k Keeper) IterateContracts(ctx sdk.Context, cb func(addr common.Address, codeHash common.Hash) (stop bool)) {
-	store := ctx.KVStore(k.storeKey)
-	iterator := storetypes.KVStorePrefixIterator(store, types.KeyPrefixCodeHash)
-
-	defer iterator.Close()
-	for ; iterator.Valid(); iterator.Next() {
-		addr := common.BytesToAddress(iterator.Key())
-		codeHash := common.BytesToHash(iterator.Value())
-
-		if cb(addr, codeHash) {
-			break
-		}
-	}
 }
 
 // GetCode loads contract code from database, implements `statedb.Keeper` interface.
@@ -144,20 +108,24 @@ func (k *Keeper) SetBalance(ctx sdk.Context, addr common.Address, amount *big.In
 // SetAccount updates nonce/balance/codeHash together.
 func (k *Keeper) SetAccount(ctx sdk.Context, addr common.Address, account statedb.Account) error {
 	// update account
-	acct := k.accountKeeper.GetAccount(ctx, addr.Bytes())
+	cosmosAddr := sdk.AccAddress(addr.Bytes())
+	acct := k.accountKeeper.GetAccount(ctx, cosmosAddr)
 	if acct == nil {
-		acct = k.accountKeeper.NewAccountWithAddress(ctx, addr.Bytes())
+		acct = k.accountKeeper.NewAccountWithAddress(ctx, cosmosAddr)
 	}
 
 	if err := acct.SetSequence(account.Nonce); err != nil {
 		return err
 	}
 
-	if types.IsEmptyCodeHash(account.CodeHash) {
-		k.DeleteCodeHash(ctx, addr)
-	} else {
-		k.SetCodeHash(ctx, addr.Bytes(), account.CodeHash)
+	codeHash := common.BytesToHash(account.CodeHash)
+
+	if ethAcct, ok := acct.(evmostypes.EthAccountI); ok {
+		if err := ethAcct.SetCodeHash(codeHash); err != nil {
+			return err
+		}
 	}
+
 	k.accountKeeper.SetAccount(ctx, acct)
 
 	if err := k.SetBalance(ctx, addr, account.Balance); err != nil {
@@ -168,80 +136,43 @@ func (k *Keeper) SetAccount(ctx sdk.Context, addr common.Address, account stated
 		"account updated",
 		"ethereum-address", addr.Hex(),
 		"nonce", account.Nonce,
-		"codeHash", common.BytesToHash(account.CodeHash).Hex(),
+		"codeHash", codeHash.Hex(),
 		"balance", account.Balance,
 	)
 	return nil
 }
 
-// SetState update contract storage.
+// SetState update contract storage, delete if value is empty.
 func (k *Keeper) SetState(ctx sdk.Context, addr common.Address, key common.Hash, value []byte) {
 	store := prefix.NewStore(ctx.KVStore(k.storeKey), types.AddressStoragePrefix(addr))
-	store.Set(key.Bytes(), value)
-
+	action := "updated"
+	if len(value) == 0 {
+		store.Delete(key.Bytes())
+		action = "deleted"
+	} else {
+		store.Set(key.Bytes(), value)
+	}
 	k.Logger(ctx).Debug(
-		"state updated",
+		fmt.Sprintf("state %s", action),
 		"ethereum-address", addr.Hex(),
 		"key", key.Hex(),
 	)
 }
 
-// DeleteState deletes the entry for the given key in the contract storage
-// at the defined contract address.
-func (k *Keeper) DeleteState(ctx sdk.Context, addr common.Address, key common.Hash) {
-	store := prefix.NewStore(ctx.KVStore(k.storeKey), types.AddressStoragePrefix(addr))
-	store.Delete(key.Bytes())
-
-	k.Logger(ctx).Debug(
-		"state deleted",
-		"ethereum-address", addr.Hex(),
-		"key", key.Hex(),
-	)
-}
-
-// SetCodeHash sets the code hash for the given contract address.
-func (k *Keeper) SetCodeHash(ctx sdk.Context, addrBytes, hashBytes []byte) {
-	store := prefix.NewStore(ctx.KVStore(k.storeKey), types.KeyPrefixCodeHash)
-	store.Set(addrBytes, hashBytes)
-
-	k.Logger(ctx).Debug(
-		"code hash updated",
-		"address", common.BytesToAddress(addrBytes).Hex(),
-		"code hash", common.BytesToHash(hashBytes).Hex(),
-	)
-}
-
-// DeleteCodeHash deletes the code hash for the given contract address from the store.
-func (k *Keeper) DeleteCodeHash(ctx sdk.Context, addr common.Address) {
-	store := prefix.NewStore(ctx.KVStore(k.storeKey), types.KeyPrefixCodeHash)
-	store.Delete(addr.Bytes())
-
-	k.Logger(ctx).Debug(
-		"code hash deleted",
-		"address", addr.Hex(),
-	)
-}
-
-// SetCode sets the given contract code bytes for the corresponding code hash bytes key
-// in the code store.
+// SetCode set contract code, delete if code is empty.
 func (k *Keeper) SetCode(ctx sdk.Context, codeHash, code []byte) {
 	store := prefix.NewStore(ctx.KVStore(k.storeKey), types.KeyPrefixCode)
-	store.Set(codeHash, code)
 
+	// store or delete code
+	action := "updated"
+	if len(code) == 0 {
+		store.Delete(codeHash)
+		action = "deleted"
+	} else {
+		store.Set(codeHash, code)
+	}
 	k.Logger(ctx).Debug(
-		"code updated",
-		"code-hash", common.BytesToHash(codeHash).Hex(),
-	)
-}
-
-// DeleteCode deletes the contract code for the given code hash bytes in
-// the corresponding store.
-func (k *Keeper) DeleteCode(ctx sdk.Context, codeHash []byte) {
-	store := prefix.NewStore(ctx.KVStore(k.storeKey), types.KeyPrefixCode)
-	store.Delete(codeHash)
-
-	k.Logger(ctx).Debug(
-		"code deleted",
+		fmt.Sprintf("code %s", action),
 		"code-hash", common.BytesToHash(codeHash).Hex(),
 	)
 }
@@ -250,7 +181,6 @@ func (k *Keeper) DeleteCode(ctx sdk.Context, codeHash []byte) {
 // - clear balance
 // - remove code
 // - remove states
-// - remove the code hash
 // - remove auth account
 func (k *Keeper) DeleteAccount(ctx sdk.Context, addr common.Address) error {
 	cosmosAddr := sdk.AccAddress(addr.Bytes())
@@ -259,9 +189,10 @@ func (k *Keeper) DeleteAccount(ctx sdk.Context, addr common.Address) error {
 		return nil
 	}
 
-	// NOTE: only Ethereum contracts can be self-destructed
-	if !k.IsContract(ctx, addr) {
-		return errors.New("only smart contracts can be self-destructed")
+	// NOTE: only Ethereum accounts (contracts) can be selfdestructed
+	_, ok := acct.(evmostypes.EthAccountI)
+	if !ok {
+		return errorsmod.Wrapf(types.ErrInvalidAccount, "type %T, address %s", acct, addr)
 	}
 
 	// clear balance
@@ -271,12 +202,9 @@ func (k *Keeper) DeleteAccount(ctx sdk.Context, addr common.Address) error {
 
 	// clear storage
 	k.ForEachStorage(ctx, addr, func(key, _ common.Hash) bool {
-		k.DeleteState(ctx, addr, key)
+		k.SetState(ctx, addr, key, nil)
 		return true
 	})
-
-	// clear code hash
-	k.DeleteCodeHash(ctx, addr)
 
 	// remove auth account
 	k.accountKeeper.RemoveAccount(ctx, acct)