+++ x/evm/statedb/mock_test.go
@@ -1,6 +1,7 @@
 package statedb_test
 
 import (
+	"bytes"
 	"errors"
 	"maps"
 	"math/big"
@@ -9,7 +10,6 @@ import (
 	"github.com/ethereum/go-ethereum/common"
 	"github.com/ethereum/go-ethereum/crypto"
 	"github.com/evmos/evmos/v19/x/evm/statedb"
-	"github.com/evmos/evmos/v19/x/evm/types"
 )
 
 var (
@@ -78,22 +78,16 @@ func (k MockKeeper) SetAccount(_ sdk.Context, addr common.Address, account state
 
 func (k MockKeeper) SetState(_ sdk.Context, addr common.Address, key common.Hash, value []byte) {
 	if acct, ok := k.accounts[addr]; ok {
-		acct.states[key] = common.BytesToHash(value)
-	}
-}
-
-func (k MockKeeper) DeleteState(_ sdk.Context, addr common.Address, key common.Hash) {
-	if acct, ok := k.accounts[addr]; ok {
-		delete(acct.states, key)
+		if len(value) == 0 {
+			delete(acct.states, key)
+		} else {
+			acct.states[key] = common.BytesToHash(value)
+		}
 	}
 }
 
 func (k MockKeeper) SetCode(_ sdk.Context, codeHash []byte, code []byte) {
 	k.codes[common.BytesToHash(codeHash)] = code
-}
-
-func (k MockKeeper) DeleteCode(_ sdk.Context, codeHash []byte) {
-	delete(k.codes, common.BytesToHash(codeHash))
 }
 
 func (k MockKeeper) DeleteAccount(_ sdk.Context, addr common.Address) error {
@@ -102,7 +96,7 @@ func (k MockKeeper) DeleteAccount(_ sdk.Context, addr common.Address) error {
 	}
 	old := k.accounts[addr]
 	delete(k.accounts, addr)
-	if !types.IsEmptyCodeHash(old.account.CodeHash) {
+	if !bytes.Equal(old.account.CodeHash, emptyCodeHash) {
 		delete(k.codes, common.BytesToHash(old.account.CodeHash))
 	}
 	return nil